+++ rope/contrib/autoimport/sqlite.py
@@ -16,7 +16,10 @@ from pathlib import Path
 from threading import local
 from typing import Generator, Iterable, Iterator, List, Optional, Set, Tuple
 
+from packaging.requirements import Requirement
+
 from rope.base import exceptions, libutils, resourceobserver, taskhandle, versioning
+from rope.base.prefs import AutoimportPrefs
 from rope.base.project import Project
 from rope.base.resources import Resource
 from rope.contrib.autoimport import models
@@ -53,18 +56,36 @@ def get_future_names(
 
 
 def filter_packages(
-    packages: Iterable[Package], underlined: bool, existing: List[str]
+    packages: Iterable[Package],
+    underlined: bool,
+    existing: List[str],
+    dependencies: Optional[List[Requirement]],
 ) -> Iterable[Package]:
     """Filter list of packages to parse."""
+    parsed_deps = (
+        [dep.name for dep in dependencies] if dependencies is not None else None
+    )
+
+    def is_dep(package) -> bool:
+        return (
+            parsed_deps is None
+            or package.name in parsed_deps
+            or package.source is not Source.SITE_PACKAGE
+        )
+
     if underlined:
 
         def filter_package(package: Package) -> bool:
-            return package.name not in existing
+            return package.name not in existing and is_dep(package)
 
     else:
 
         def filter_package(package: Package) -> bool:
-            return package.name not in existing and not package.name.startswith("_")
+            return (
+                package.name not in existing
+                and not package.name.startswith("_")
+                and is_dep(package)
+            )
 
     return filter(filter_package, packages)
 
@@ -81,16 +102,15 @@ class AutoImport:
     """
 
     connection: sqlite3.Connection
-    memory: bool
     project: Project
     project_package: Package
-    underlined: bool
+    prefs: AutoimportPrefs
 
     def __init__(
         self,
         project: Project,
         observe: bool = True,
-        underlined: bool = False,
+        underlined: Optional[bool] = None,
         memory: bool = _deprecated_default,
     ):
         """Construct an AutoImport object.
@@ -113,25 +133,29 @@ class AutoImport:
                 autoimport = AutoImport(..., memory=True)
         """
         self.project = project
+        self.prefs = self.project.prefs.autoimport
         project_package = get_package_tuple(project.root.pathlib, project)
         assert project_package is not None
         assert project_package.path is not None
+        if underlined is not None:
+            self.prefs.underlined = underlined
         self.project_package = project_package
-        self.underlined = underlined
-        self.memory = memory
         if memory is _deprecated_default:
-            self.memory = True
-            warnings.warn(
-                "The default value for `AutoImport(memory)` argument will "
-                "change to use an on-disk database by default in the future. "
-                "If you want to use an in-memory database, you need to pass "
-                "`AutoImport(memory=True)` explicitly.",
-                DeprecationWarning,
-            )
+            if self.prefs.memory is None:
+                self.prefs.memory = True
+                warnings.warn(
+                    "The default value for `AutoImport(memory)` argument will "
+                    "change to use an on-disk database by default in the future. "
+                    "If you want to use an in-memory database, you need to pass "
+                    "`AutoImport(memory=True)` explicitly or set it in the config file.",
+                    DeprecationWarning,
+                )
+        else:
+            self.prefs.memory = memory
         self.thread_local = local()
         self.connection = self.create_database_connection(
             project=project,
-            memory=memory,
+            memory=self.prefs.memory,
         )
         self._setup_db()
         if observe:
@@ -389,12 +413,13 @@ class AutoImport:
         """
         Generate global name cache for external modules listed in `modules`.
 
-        If no modules are provided, it will generate a cache for every module available.
+        If modules is not specified, uses PEP 621 metadata.
+        If modules aren't specified and PEP 621 is not present, caches every package
         This method searches in your sys.path and configured python folders.
         Do not use this for generating your own project's internal names,
         use generate_resource_cache for that instead.
         """
-        underlined = self.underlined if underlined is None else underlined
+        underlined = self.prefs.underlined if underlined is None else underlined
 
         packages: List[Package] = (
             self._get_available_packages()
@@ -403,12 +428,16 @@ class AutoImport:
         )
 
         existing = self._get_packages_from_cache()
-        packages = list(filter_packages(packages, underlined, existing))
-        if not packages:
+        packages = list(
+            filter_packages(
+                packages, underlined, existing, self.project.prefs.dependencies
+            )
+        )
+        if len(packages) == 0:
             return
         self._add_packages(packages)
         job_set = task_handle.create_jobset("Generating autoimport cache", 0)
-        if single_thread:
+        if single_thread or not self.prefs.parallel:
             for package in packages:
                 for module in get_files(package, underlined):
                     job_set.started_job(module.modname)
@@ -512,7 +541,7 @@ class AutoImport:
         self, resource: Resource, underlined: bool = False, commit: bool = True
     ):
         """Update the cache for global names in `resource`."""
-        underlined = underlined if underlined else self.underlined
+        underlined = underlined if underlined else self.prefs.underlined
         module = self._resource_to_module(resource, underlined)
         self._del_if_exist(module_name=module.modname, commit=False)
         for name in get_names(module, self.project_package):
@@ -537,7 +566,11 @@ class AutoImport:
 
     def _get_python_folders(self) -> List[Path]:
         def filter_folders(folder: Path) -> bool:
-            return folder.is_dir() and folder.as_posix() != "/usr/bin"
+            return (
+                folder.is_dir()
+                and folder.as_posix() != "/usr/bin"
+                and str(folder) != self.project.address
+            )
 
         folders = self.project.get_python_path_folders()
         folder_paths = filter(filter_folders, map(Path, folders))
@@ -623,7 +656,7 @@ class AutoImport:
         self, resource: Resource, underlined: bool = False
     ) -> ModuleFile:
         assert self.project_package.path
-        underlined = underlined if underlined else self.underlined
+        underlined = underlined if underlined else self.prefs.underlined
         resource_path: Path = resource.pathlib
         # The project doesn't need its name added to the path,
         # since the standard python file layout accounts for that