+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # **Upcoming release**
-
+- #516 Autoimport Now automatically detects project dependencies and can read TOML configuration
 - #733 skip directories with perm error when building autoimport index (@MrBago)
 - #722, #723 Remove site-packages from packages search tree (@tkrabel)
 - #738 Implement os.PathLike on Resource (@lieryan)